+++ pkg/client/connector/traffic_manager.go
@@ -4,10 +4,12 @@ import (
 	"bufio"
 	"context"
 	"fmt"
+	"net"
 	"os"
 	"os/user"
 	"regexp"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 
@@ -15,8 +17,10 @@ import (
 	"google.golang.org/grpc"
 	errors2 "k8s.io/apimachinery/pkg/api/errors"
 
+	"github.com/datawire/dlib/dexec"
 	"github.com/datawire/dlib/dgroup"
 	"github.com/datawire/dlib/dlog"
+	"github.com/datawire/dlib/dtime"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/connector"
 	"github.com/telepresenceio/telepresence/rpc/v2/manager"
 	"github.com/telepresenceio/telepresence/v2/pkg/client"
@@ -309,6 +313,7 @@ func (tm *trafficManager) setStatus(ctx context.Context, r *rpc.ConnectInfo) {
 	if tm == nil {
 		return
 	}
+	r.BridgeOk = tm.check(ctx)
 	if tm.managerClient == nil {
 		r.Intercepts = &manager.InterceptInfoSnapshot{}
 		r.Agents = &manager.AgentInfoSnapshot{}
@@ -389,4 +394,86 @@ func (tm *trafficManager) uninstall(c context.Context, ur *rpc.UninstallRequest)
 		}
 	}
 	return result, nil
+}
+
+// check checks the status of teleproxy bridge by doing the equivalent of
+//  curl http://traffic-manager.svc:8022.
+// Note there is no namespace specified, as we are checking for bridge status in the
+// current namespace.
+func (br *trafficManager) check(c context.Context) bool {
+	if br == nil {
+		return false
+	}
+	address := fmt.Sprintf("localhost:%d", br.sshPort)
+	conn, err := net.DialTimeout("tcp", address, 15*time.Second)
+	if err != nil {
+		dlog.Errorf(c, "fail to establish tcp connection to %s: %v", address, err)
+		return false
+	}
+	defer conn.Close()
+
+	msg, _, err := bufio.NewReader(conn).ReadLine()
+	if err != nil {
+		dlog.Errorf(c, "tcp read: %v", err)
+		return false
+	}
+	if !strings.Contains(string(msg), "SSH") {
+		dlog.Errorf(c, "expected SSH prompt, got: %v", string(msg))
+		return false
+	}
+	return true
+}
+
+// sshPortForward synchronously runs an `ssh` process with the given port-forward args.  It retries
+// for all errors; it only returns when the context is canceled.  For that reason, it doesn't return
+// an error; it just always retries.
+func (tm *trafficManager) sshPortForward(ctx context.Context, pfArgs ...string) {
+	// XXX: probably need some kind of keepalive check for ssh, first
+	// curl after wakeup seems to trigger detection of death
+	sshArgs := append(append([]string{"ssh"}, pfArgs...), []string{
+		"-F", "none", // don't load the user's config file
+
+		// connection settings
+		"-C", // compression
+		"-oConnectTimeout=10",
+		"-oStrictHostKeyChecking=no",     // don't bother checking the host key...
+		"-oUserKnownHostsFile=/dev/null", // and since we're not checking it, don't bother remembering it either
+
+		// port-forward settings
+		"-N", // no remote command; just connect and forward ports
+		"-oExitOnForwardFailure=yes",
+
+		// where to connect to
+		"-p", strconv.Itoa(int(tm.sshPort)),
+		"telepresence@localhost",
+	}...)
+
+	// Do NOT use client.Retry; this has slightly more domain-specific knowledge regarding the
+	// backoff: We don't backoff if the process was "long-lived".  SSH connections just
+	// sometimes die; we should retry those immediately; we only want to back off when it looks
+	// like there's a problem *establishing* the connection.  So we use process-lifetime as a
+	// proxy for whether a connection was established or not.
+	backoff := 100 * time.Millisecond
+	for ctx.Err() == nil {
+		start := time.Now()
+		err := dexec.CommandContext(ctx, sshArgs[0], sshArgs[1:]...).Run()
+		lifetime := time.Since(start)
+
+		if ctx.Err() == nil {
+			if err == nil {
+				err = errors.New("ssh process terminated successfully, but unexpectedly")
+			}
+			dlog.Errorf(ctx, "communicating with manager: %v", err)
+			if lifetime >= 20*time.Second {
+				backoff = 100 * time.Millisecond
+				dtime.SleepWithContext(ctx, backoff)
+			} else {
+				dtime.SleepWithContext(ctx, backoff)
+				backoff *= 2
+				if backoff > 3*time.Second {
+					backoff = 3 * time.Second
+				}
+			}
+		}
+	}
 }